requirements.txt
# Core ML and Deep Learning
torch==2.2.2
torchvision==0.17.2
transformers==4.52.1
datasets==3.6.0
tokenizers==0.21.1

# Scientific Computing and Data
numpy==1.26.4
pandas==2.2.3
matplotlib==3.10.3

# Configuration and Utilities
PyYAML==6.0.2
python-dotenv==1.0.1

# Experiment Tracking
wandb==0.19.11

# Development and Testing
pytest==8.4.0
black==25.1.0
flake8==7.2.0
flake8-bugbear==24.12.12

# Core Python packages (typically included but ensuring compatibility)
requests==2.32.3
tqdm==4.67.1
packaging==24.2
